+++ changelog.md
@@ -3,6 +3,7 @@
 #future
 
 #1.0.3
+- ore shrubs integration (https://minecraft.curseforge.com/projects/ore-shrubs)
 
 #1.0.2
 - moved to MMD servers